pyproject.toml
[tool.poetry]
name = "py-de-familia"
version = "0.1.3"
description = "CLI deliciosa de família"
homepage = "https://github.com/augustoliks/py-de-familia"
authors = ["augustoliks <carlos.neto.dev@gmail.com>"]
maintainers = ["augustoliks <carlos.neto.dev@gmail.com>"]
readme = "README.rst"
#license = "APAICHE-DE-FAMILIA-2.0"
documentation = "https://py-de-familia.readthedocs.io/en/latest/?badge=latest"
keywords = ["cli", "relaxar", "ai", "macho"]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
playsound = "^1.2.2"
pytest = "^6.1.2"
click_completion = "^0.5.2"
pytest-cov = "^2.10.1"
python-telegram-bot = "^13.1"
dynaconf = "^3.1.2"

[tool.poetry.scripts]
py-de-familia = "py_de_familia.clidefamilia:entrypoint_de_familia"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
